refactor: extract meta api
