+++ contract/src/lib.rs
@@ -2,11 +2,7 @@
 extern crate static_assertions;
 
 use api::{Payout, RestrictionApi, SweatApi};
-use near_contract_standards::fungible_token::{
-    events::FtBurn,
-    metadata::{FungibleTokenMetadata, FungibleTokenMetadataProvider},
-    FungibleToken,
-};
+use near_contract_standards::fungible_token::{events::FtBurn, FungibleToken};
 use near_plugins::{access_control, pause, AccessControlRole, AccessControllable, Pausable};
 use near_sdk::{
     borsh::BorshDeserialize,
@@ -23,6 +19,7 @@ mod core;
 mod defer;
 mod integration;
 mod math;
+mod meta;
 mod migration;
 
 #[derive(AccessControlRole, Deserialize, Serialize, Clone, Copy)]
@@ -122,23 +119,6 @@ impl Contract {
                 !self.is_restricted(account_id),
                 format!("The account {account_id} is restricted")
             );
-        }
-    }
-}
-
-pub const ICON: &str = "data:image/svg+xml,%3Csvg viewBox='0 0 100 100' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Crect width='100' height='100' rx='50' fill='%23FF0D75'/%3E%3Cg clip-path='url(%23clip0_283_2788)'%3E%3Cpath d='M39.4653 77.5455L19.0089 40.02L35.5411 22.2805L55.9975 59.806L39.4653 77.5455Z' stroke='white' stroke-width='10'/%3E%3Cpath d='M66.0253 77.8531L45.569 40.3276L62.1012 22.5882L82.5576 60.1136L66.0253 77.8531Z' stroke='white' stroke-width='10'/%3E%3C/g%3E%3Cdefs%3E%3CclipPath id='clip0_283_2788'%3E%3Crect width='100' height='56' fill='white' transform='translate(0 22)'/%3E%3C/clipPath%3E%3C/defs%3E%3C/svg%3E%0A";
-
-#[near]
-impl FungibleTokenMetadataProvider for Contract {
-    fn ft_metadata(&self) -> FungibleTokenMetadata {
-        FungibleTokenMetadata {
-            spec: "ft-1.0".to_string(),
-            name: "SWEAT".to_string(),
-            symbol: "SWEAT".to_string(),
-            icon: Some(String::from(ICON)),
-            reference: None,
-            reference_hash: None,
-            decimals: 18,
         }
     }
 }
